you've got to be kidding me...

+++ G007-G008/Lab4/team/testMean.cpp
@@ -21,7 +21,7 @@ bool testMean(){
 }
 
 bool testMean(std::vector<int> input_set){
-	uint input_size = input_set.size();
+	float input_size = input_set.size();
 
 	bool output = true;
